requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
jinja2>=3.1.0

# Text Processing
pyahocorasick>=2.0.0
jieba>=0.42.1
opencc-python-reimplemented>=0.1.7

# Additional utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
pyyaml>=6.0.1

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
ruff>=0.1.0
